Remove number of methods in FixturesAwareTestCase by moving generic container/kernel related to a new AbstractTestCase
